// src/proto/broadcast.ts
export interface BroadcastFrame {
  /** 请求消息信息 */
  options:
    | FrameOption
    | undefined;
  /** 业务target_path */
  targetPath: string;
  /** 业务pb内容 */
  body: Any | undefined;
}

/** google.protobuf.Any */
export interface Any {
  typeUrl: string;
  value: Uint8Array;
}

/** google.protobuf.Empty */
export interface Empty {
}

/** 请求消息信息 */
export interface FrameOption {
  /**
   * 消息id
   *   - client: 本次连接唯一的消息id，可用于回执
   *   - server: 唯一消息id，可用于上报或者回执
   */
  messageId?:
    | bigint
    | undefined;
  /**
   * frame序号
   *   - client: 客户端应该每次请求时frame seq++，会返回对应的对称req/resp
   *   - server: 服务端下行消息，只会返回默认值：0
   */
  sequence: bigint;
  /**
   * 是否进行消息回执(发出MessageAckReq)
   * downstream 上只有服务端设置为true，客户端响应
   * upstream   上只有客户端设置为true，服务端响应
   * 响应帧禁止设置is_ack，协议上禁止循环
   * 通常只有业务帧才可能设置is_ack, 因为协议栈(例如心跳、鉴权)另有响应约定
   */
  isAck?:
    | boolean
    | undefined;
  /** 业务状态码 */
  status?:
    | Status
    | undefined;
  /** 业务ack来源, 仅downstream时候由服务端填写. */
  ackOrigin?: string | undefined;
}

/** 业务状态码 */
export interface Status {
  /**
   * 业务错误码
   * 当code是[UNKNOWN = 2]时，details为业务详细的错误信息，进行proto any转换成业务码结构体
   */
  code: number;
  /** 对错误码的的解释 */
  message: string;
  /** 是调用方和被调用方约定好的额外的proto结构，包含一些详细的错误处理方案等 */
  details: Any[];
}

/** 鉴权请求，通过authorization验证绑定用户mid */
export interface AuthReq {
  /** 冷启动id，算法uuid，重新起启会变 */
  guid: string;
  /** 连接id，算法uuid，重连会变 */
  connId?:
    | string
    | undefined;
  /** 最后收到的消息id，用于过虑重连后获取未读的消息 */
  lastMsgId?: bigint | undefined;
}

/** 鉴权返回 */
export interface AuthResp {
}

/** 消息回执 */
export interface MessageAckReq {
  /** 消息id */
  ackId: bigint;
  /** ack来源，由业务指定用于埋点跟踪 */
  ackOrigin: string;
  /** 消息对应的target_path，方便业务区分和监控统计 */
  targetPath: string;
}

/**
 * 心跳请求
 * 心跳保活：成功心跳为4分45秒，重试心跳为30s，三次收不到进行重连（不超过5分45）
 */
export interface HeartbeatReq {
}

/** 心跳返回 */
export interface HeartbeatResp {
}

/** 路径订阅请求 */
export interface TargetPath {
  /** 需要订阅的target_paths */
  targetPaths: string[];
}

/** 房间操作请求 */
export interface RoomReq {
  /** {type}://{room_id} 如 video://aid/cid */
  id: string;
  event:
    | { $case: "join"; value: RoomJoinEvent }
    | { $case: "leave"; value: RoomLeaveEvent }
    | { $case: "online"; value: RoomOnlineEvent }
    | { $case: "msg"; value: RoomMessageEvent }
    | undefined;
}

/** 房间操作返回 */
export interface RoomResp {
  /** {type}://{room_id} 如 video://aid/cid */
  id: string;
  event:
    | { $case: "join"; value: RoomJoinEvent }
    | { $case: "leave"; value: RoomLeaveEvent }
    | { $case: "online"; value: RoomOnlineEvent }
    | { $case: "msg"; value: RoomMessageEvent }
    | { $case: "err"; value: RoomErrorEvent }
    | undefined;
}

export interface RoomJoinEvent {
}

export interface RoomLeaveEvent {
}

export interface RoomOnlineEvent {
  online?: number | undefined;
}

export interface RoomMessageEvent {
  targetPath: string;
  body: Any | undefined;
}

export interface RoomErrorEvent {
  status: Status | undefined;
}

function createBaseBroadcastFrame(): BroadcastFrame {
  return { options: undefined, targetPath: "", body: undefined };
}

export const BroadcastFrame: MessageFns<BroadcastFrame> = {
  encode(message: BroadcastFrame, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.options !== undefined) {
      FrameOption.encode(message.options, writer.uint32(10).fork()).join();
    }
    if (message.targetPath !== "") {
      writer.uint32(18).string(message.targetPath);
    }
    if (message.body !== undefined) {
      Any.encode(message.body, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BroadcastFrame {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBroadcastFrame();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.options = FrameOption.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.targetPath = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.body = Any.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseAny(): Any {
  return { typeUrl: "", value: new Uint8Array(0) };
}

export const Any: MessageFns<Any> = {
  encode(message: Any, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.typeUrl !== "") {
      writer.uint32(10).string(message.typeUrl);
    }
    if (message.value.length !== 0) {
      writer.uint32(18).bytes(message.value);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Any {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAny();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.typeUrl = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.value = reader.bytes();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const AuthReq: MessageFns<AuthReq> = {
  encode(message: AuthReq, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.guid !== "") {
      writer.uint32(10).string(message.guid);
    }
    if (message.connId !== undefined) {
      writer.uint32(18).string(message.connId);
    }
    if (message.lastMsgId !== undefined) {
      if (BigInt.asIntN(64, message.lastMsgId) !== message.lastMsgId) {
        throw new globalThis.Error("value provided for field message.lastMsgId of type int64 too large");
      }
      writer.uint32(24).int64(message.lastMsgId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AuthReq {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAuthReq();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.guid = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.connId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.lastMsgId = reader.int64() as bigint;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const MessageAckReq: MessageFns<MessageAckReq> = {
  encode(message: MessageAckReq, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.ackId !== 0n) {
      if (BigInt.asIntN(64, message.ackId) !== message.ackId) {
        throw new globalThis.Error("value provided for field message.ackId of type int64 too large");
      }
      writer.uint32(8).int64(message.ackId);
    }
    if (message.ackOrigin !== "") {
      writer.uint32(18).string(message.ackOrigin);
    }
    if (message.targetPath !== "") {
      writer.uint32(26).string(message.targetPath);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MessageAckReq {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMessageAckReq();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.ackId = reader.int64() as bigint;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.ackOrigin = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.targetPath = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseHeartbeatReq(): HeartbeatReq {
  return {};
}

export const HeartbeatReq: MessageFns<HeartbeatReq> = {
  encode(_: HeartbeatReq, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): HeartbeatReq {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseHeartbeatReq();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const TargetPath: MessageFns<TargetPath> = {
  encode(message: TargetPath, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.targetPaths) {
      writer.uint32(10).string(v!);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TargetPath {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTargetPath();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.targetPaths.push(reader.string());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseRoomReq(): RoomReq {
  return { id: "", event: undefined };
}

export const RoomReq: MessageFns<RoomReq> = {
  encode(message: RoomReq, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    switch (message.event?.$case) {
      case "join":
        RoomJoinEvent.encode(message.event.value, writer.uint32(18).fork()).join();
        break;
      case "leave":
        RoomLeaveEvent.encode(message.event.value, writer.uint32(26).fork()).join();
        break;
      case "online":
        RoomOnlineEvent.encode(message.event.value, writer.uint32(34).fork()).join();
        break;
      case "msg":
        RoomMessageEvent.encode(message.event.value, writer.uint32(42).fork()).join();
        break;
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RoomReq {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRoomReq();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.event = { $case: "join", value: RoomJoinEvent.decode(reader, reader.uint32()) };
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.event = { $case: "leave", value: RoomLeaveEvent.decode(reader, reader.uint32()) };
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.event = { $case: "online", value: RoomOnlineEvent.decode(reader, reader.uint32()) };
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.event = { $case: "msg", value: RoomMessageEvent.decode(reader, reader.uint32()) };
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
}